feat: dynamic configuration
