+++ src/main/java/com/festp/styledmessage/StyledMessageBuilderFactory.java
@@ -1,0 +1,6 @@
+package com.festp.styledmessage;
+
+public interface StyledMessageBuilderFactory
+{
+	public StyledMessageBuilder create();
+}
